+++ src/model/warehouse/PurchaseWarehouse.java
@@ -10,7 +10,7 @@ public class PurchaseWarehouse extends Warehouse{
 
 	public void sell() {
 		//TODO: Falta sumar el 50% de las máquinas en buen estado
-		budget.increment((float)(0.8 * ground.getPrice()));
+		budget.increment((int)(0.8 * ground.getPrice()));
 	}
 
 	public void updateDay() {
